RAG_tutorials/vision_rag/pyproject.toml
[project]
name = "vision-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.0",
]
